@@ -1,8 +1,8 @@
-package com.besign.contests.hashCode.persistence;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MatrixElementRepository extends CrudRepository<MatrixElement, MatrixId> {
-}
+//package com.besign.contests.hashCode.persistence;
+//
+//import org.springframework.data.repository.CrudRepository;
+//import org.springframework.stereotype.Repository;
+//
+//@Repository
+//public interface MatrixElementRepository extends CrudRepository<MatrixElement, MatrixId> {
+//}
